fix(client): fixed agent hiding/unhiding again
after the the rewrite of HcAgent object to use the IHcAgent interface
the ::hide() and ::unhide() has been rewritten into a single ::setHidden
call. Just that this call also been used for actionShowHidden which
reverts back the selected hidden agent back to be visiable again.
re-introduced ::hide() and ::unhide() just for this. setHidden is to be
used to set the agent hidden state.